feat: 게시판 구현
BoardInterceptor, deletePost, updatePost - 현재 세션 기반 유저 id 검사

+++ server/src/main/java/com/mayonedev/battle/controller/BoardController.java
@@ -10,16 +10,18 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.mayonedev.battle.entity.Board;
+import com.mayonedev.battle.entity.User;
 import com.mayonedev.battle.service.BoardService;
 
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -40,30 +42,23 @@ public class BoardController {
         return ResponseEntity.ok(boards);
     }
 
-    // TODO 자기 게시글 보기
-    // @PostMapping("/")
-    // @Operation(summary = "게시글 상세 조회", description = "userId 를 이용해 게시글을 상세 조회 합니다.")
-    // public ResponseEntity<List<Board>> getPostsByUserId(
-    //     @Parameter(description = "사용자 ID") @PathVariable long id){
-
-    //         List<Board> board = boardService.getPostByUserId(id);
-
-    //         if(board.isEmpty()){
-    //             return ResponseEntity.notFound().build(); // TODO: Q. 이게 무슨 코드일까...
-    //         }
-
-    //         return ResponseEntity.ok(board);
-    // }
-
     // 게시글 상세 조회
     @GetMapping("/post/{id}")
     @Operation(summary = "게시글 상세 조회", description = "postId 를 이용해 게시글을 상세 조회 합니다.")
     public ResponseEntity<Board> getPostByPostId( @PathVariable int id) throws Exception{
-        Board board = boardService.selectByPostId(id);
+    	 
+    	 boardService.increaseViewCount(id); //조회수 카운트
+    	 //근데 얘도 sql 에서 update 인건 똑같은데 왜 getmapping이지.. 게시글 수정이랑 뭐가 다른 걸까..
+    	 
+    	 Board board = boardService.selectByPostId(id);
+    	 
+    	 System.out.println(board);
+        
+        Map<String, Object> map = Map.of( "resmsg", "게시글 상세 조회", "resvalue", board);
         return ResponseEntity.ok(board);
     }
 
-    // 게시글 작성 - 유저 id?
+    // 게시글 작성
     @PostMapping("/post")
     @Operation(summary = "게시글 작성", description = "게시글을 작성 합니다.")
     public ResponseEntity<?> craeteBoard(@RequestBody Board b) throws Exception{
@@ -81,16 +76,23 @@ public class BoardController {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(err);
         }
     }
-    // 게시글 수정 - 자기 게시글만 수정 가능하게끔...
+    // 게시글 수정 - 현재 세션 기반 유저 검사 
     @PatchMapping("/post/{id}")
-    @Operation(summary = "게시글 수정", description = "게시글을 수정합니다.")
-    public ResponseEntity<?> updatePost(@PathVariable int id, @RequestBody Board b) throws Exception{
+    @Operation(summary = "게시글 수정", description = "본인 게시글을 수정합니다.")
+    public ResponseEntity<?> updatePost(@PathVariable int id, @RequestBody Board b, HttpSession session) throws Exception{
        try {
-            boardService.updatePost(b);
-            Map<String, Object> map = Map.of("resmsg", "게시글이 수정되었습니다.", "resvalue", b);
-            return ResponseEntity
-                    .status(HttpStatus.CREATED)
-                    .body(map);
+    	    User loginUser = (User) session.getAttribute("loginUser");
+    	    if(loginUser.getId()==b.getId()) {
+    	    	boardService.updatePost(b);
+                Map<String, Object> map = Map.of("resmsg", "게시글이 수정되었습니다.", "resvalue", b);
+                return ResponseEntity
+                        .status(HttpStatus.CREATED)
+                        .body(map);
+    	    }else {
+    	    	Map<String, Object> err = Map.of("resmsg", "게시글 수정에 실패했습니다",
+                        "resvalue", "게시글 작성자만 게시글을 수정할 수 있습니다." );
+    	    		return ResponseEntity.status(HttpStatus.FORBIDDEN).body(err);
+    	    }
 
         } catch (RuntimeException e) {
             Map<String, Object> err = Map.of("resmsg", "게시글 수정에 실패했습니다",
@@ -100,16 +102,25 @@ public class BoardController {
 
     }
 
-    // 게시글 삭제
+    // 게시글 삭제 - 현재 세션 기반 유저 검사 
     @DeleteMapping("/post/{id}")
-    @Operation(summary = "게시글 삭제", description = "게시글을 삭제합니다.")
-    public ResponseEntity<?> deletePost( @PathVariable int id) throws Exception{
+    @Operation(summary = "게시글 삭제", description = "본인 게시글을 삭제합니다.")
+    public ResponseEntity<?> deletePost(@PathVariable int id, HttpSession session) throws Exception{
         try {
-            boardService.deletePost(id);
-            Map<String, Object> map = Map.of("resmsg", "게시글이 삭제되었습니다.", "resvalue", id);
-            return ResponseEntity
-                    .status(HttpStatus.CREATED)
-                    .body(map);
+        	User loginUser = (User) session.getAttribute("loginUser");
+        	if(loginUser.getId() == id) {
+        		boardService.deletePost(id);
+                Map<String, Object> map = Map.of("resmsg", "게시글이 삭제되었습니다.", "resvalue", id);
+                return ResponseEntity
+                        .status(HttpStatus.CREATED)
+                        .body(map);
+        	}else {
+        		Map<String, Object> err = Map.of("resmsg", "게시글 삭제에 실패했습니다.", "resvalue", "본인이 작성한 게시글만 삭제가 가능합니다.");
+        		return ResponseEntity
+        				.status(HttpStatus.FORBIDDEN).body(err);
+        	}
+            
+           
         } catch (RuntimeException e) {
                 Map<String, Object> err = Map.of("resmsg", "게시글 삭제에 실패했습니다",
                                                  "resvalue", e.getMessage() );
